lets try this

+++ workflow/rules/5.concoct.smk
@@ -131,6 +131,7 @@ rule rename_concoctBins:
         for sample in {params.sample}; do
             src_dir={dir_binning}/${{sample}}_concoct/bins
             for f in "$src_dir"/concoct_*.fa; do
+                [ -e "$f" ] || continue   # <-- critical guard
                 cp "$f" "{params.outdir}/.
             done
         done
